packages/spectacles/requirements.txt
# Spectacles - Dependencies
# SOC2/HIPAA Compliant Automation Platform

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0

# Browser Automation
playwright>=1.40.0

# Desktop Automation (VM only - not installed on Cloud Run)
# These are optional dependencies for desktop mode
mss>=9.0.0                    # Fast cross-platform screen capture
pyautogui>=0.9.54             # Mouse and keyboard automation
pygetwindow>=0.0.9            # Window detection and management
easyocr>=1.7.0                # OCR for native app text extraction

# File System Monitoring
watchdog>=3.0.0               # File system event monitoring

# Slack Integration
slack-bolt>=1.18.0
slack-sdk>=3.27.0

# AI/VLM (Gemini 2.0 Flash)
google-generativeai>=0.8.0

# MCP Server
mcp>=1.0.0

# Database
aiosqlite>=0.20.0

# Security & Encryption
cryptography>=42.0
google-cloud-secret-manager>=2.20.0

# Image Processing (PII Blur)
Pillow>=10.0.0

# Vector Database (Qdrant Embedded - Long-term Memory)
qdrant-client==1.7.0
sentence-transformers==2.3.1

# HTTP Client
httpx>=0.27.0
aiohttp>=3.9.0

# Logging
structlog>=24.0.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Type Checking
mypy>=1.8.0

# Linting
ruff>=0.1.0
